Remove orders route
