+++ crates/e2e/tests/e2e/services.rs
@@ -259,7 +259,7 @@ impl OrderbookServices {
             Duration::from_secs(120),
             Duration::MAX,
             SignatureConfiguration::all(),
-            bad_token_detector.clone(),
+            bad_token_detector,
             quoter.clone(),
             balance_fetcher,
             signature_validator,
@@ -268,7 +268,6 @@ impl OrderbookServices {
             contracts.domain_separator,
             contracts.gp_settlement.address(),
             db.clone(),
-            bad_token_detector,
             solvable_orders_cache.clone(),
             Duration::from_secs(600),
             order_validator.clone(),
