+++ crates/orderbook/openapi.yml
@@ -47,75 +47,6 @@ paths:
           application/json:
             schema:
               $ref: "#/components/schemas/OrderCreation"
-    get:
-      summary: Get existing orders.
-      description: |
-        By default all currently valid orders are returned. The set of returned orders can be
-        reduced by setting owner, sell token, buy token filters. It can be increased by disabling
-        different order validity exclusion criteria.
-        At least one of owner, sellToken, buyToken has to be set.
-      parameters:
-        - name: owner
-          in: query
-          schema:
-            $ref: "#/components/schemas/Address"
-          required: false
-        - name: sellToken
-          in: query
-          schema:
-            $ref: "#/components/schemas/Address"
-          required: false
-        - name: buyToken
-          in: query
-          schema:
-            $ref: "#/components/schemas/Address"
-          required: false
-        - name: includeFullyExecuted
-          in: query
-          description: Should fully executed orders be returned?
-          schema:
-            type: boolean
-            default: false
-        - name: includeInvalidated
-          in: query
-          description: Should orders that have been invalidated be returned?
-          schema:
-            type: boolean
-            default: false
-        - name: includeInsufficientBalance
-          in: query
-          description: Should fill or kill orders that are not sufficiently funded be included?
-          schema:
-            type: boolean
-            default: false
-        - name: includePresignaturePending
-          in: query
-          description: Should pre-sign orders waiting for the on-chain presignature be included?
-          schema:
-            type: boolean
-            default: false
-        - name: includeUnsupportedTokens
-          in: query
-          description: Should the orders containing unsupported tokens be included?
-          schema:
-            type: boolean
-            default: false
-        - name: minValidTo
-          in: query
-          description: |
-            Minimum valid_to timestamp for included orders.
-            The default is the current time.
-          schema:
-            type: integer
-      responses:
-        200:
-          description: existing orders
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: "#/components/schemas/Order"
   /api/v1/orders/{UID}:
     get:
       summary: Get existing order from UID.
